requirements.txt
# Core Framework
streamlit>=1.28.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# Machine Learning & NLP
sentence-transformers>=2.2.2
torch>=2.0.0
torchvision>=0.15.0
spacy>=3.6.0
numpy>=1.24.0
scikit-learn>=1.3.0
medcat
nltk
# Vector Database
faiss-cpu>=1.7.4

# Web Search
duckduckgo-search>=3.9.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
json5>=0.9.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.7.0
pydantic>=2.0.0

